+++ Frontend/src/pages/index.ts
@@ -1,5 +1,3 @@
 export { default as IntroPage } from './IntroPage';
 export { default as Homepage } from './Homepage';
 export { default as SurveyPage } from './SurveyPage/SurveyPage';
-
-export { default as SurveyContextProvider } from './SurveyPage/Survey.context';
